+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+# 0.1.9 - 2019-07-04
+
 * Updated to support `futures-preview` 0.3.0-alpha.17.
 
 * Updated minimum `derive_utils` version to 0.8.0.